+++ oiosaml/src/test/java/dk/gov/oio/saml/TestSuite.java
@@ -6,12 +6,13 @@ import org.junit.runner.RunWith;
 
 @RunWith(JUnitPlatform.class)
 @SelectPackages( {
-    "dk.gov.oio.saml.audit",
     "dk.gov.oio.saml.filter",
     "dk.gov.oio.saml.oiobpp",
     "dk.gov.oio.saml.service",
     "dk.gov.oio.saml.service.validation",
     "dk.gov.oio.saml.servlet",
+    "dk.gov.oio.saml.util",
+    "dk.gov.oio.saml.audit",
     "dk.gov.oio.saml.session",
     "dk.gov.oio.saml.session.database",
     "dk.gov.oio.saml.session.inmenory"