+++ RELEASE_NOTES.md
@@ -5,6 +5,7 @@
 - NLRFIM-121: Session handling and storage in Java OIO SAML 3
 - NLRFIM-123: Replay protection in Java OIO SAML 3
 - CVE-2021-45105, CVE-2021-44832: Update log4j-core to version 2.17.1
+- NLRFIM-139: Absolute path OIOSAML configuration
 
 ## 3.1.0
 - CVE-2021-44228: Update log4j-core to version 2.16.0